moved splinediff, polydiff, and savgoldiff to a new module called polynomial_fit, leaving method stubs behind so everything is backwards compatible. The tests still pass without modification, importing from the old places.
