+++ docs/source/polynomial_fit.rst
@@ -1,0 +1,5 @@
+polynomial_fit
+==============
+
+.. automodule:: pynumdiff.polynomial_fit
+    :members:
