+++ docs/source/utils/evaluate.rst
@@ -1,7 +1,5 @@
 evaluate
 ========
 
-.. currentmodule:: pynumdiff.utils.evaluate
-
 .. automodule:: pynumdiff.utils.evaluate
     :members:
