+++ docs/source/utils/simulate.rst
@@ -1,7 +1,5 @@
 simulate
 ========
 
-.. currentmodule:: pynumdiff.utils.simulate
-
 .. automodule:: pynumdiff.utils.simulate
     :members:
