+++ docs/source/utils/utility.rst
@@ -1,7 +1,5 @@
 utility
 =======
 
-.. currentmodule:: pynumdiff.utils.utility
-
 .. automodule:: pynumdiff.utils.utility
     :members:
