+++ pynumdiff/__init__.py
@@ -1,11 +1,12 @@
 """Import useful functions from all modules
 """
 from pynumdiff._version import __version__
-from pynumdiff.finite_difference import first_order, second_order, fourth_order
+from pynumdiff.finite_difference import finite_difference, first_order, second_order, fourth_order
 from pynumdiff.smooth_finite_difference import mediandiff, meandiff, gaussiandiff,\
-    friedrichsdiff, butterdiff, splinediff
-from pynumdiff.total_variation_regularization import iterative_velocity, velocity,\
-    acceleration, jerk, smooth_acceleration, jerk_sliding
-from pynumdiff.linear_model import lineardiff, polydiff, spectraldiff, savgoldiff
-from pynumdiff.kalman_smooth import constant_velocity, constant_acceleration, constant_jerk,\
+    friedrichsdiff, butterdiff
+from pynumdiff.polynomial_fit import splinediff, polydiff, savgoldiff
+from pynumdiff.total_variation_regularization import tvrdiff, velocity, acceleration, jerk,\
+    iterative_velocity, smooth_acceleration, jerk_sliding
+from pynumdiff.linear_model import lineardiff, spectraldiff
+from pynumdiff.kalman_smooth import rts_const_deriv, constant_velocity, constant_acceleration, constant_jerk,\
     known_dynamics
