+++ pynumdiff/linear_model/__init__.py
@@ -10,4 +10,6 @@ except:
 
 from ._linear_model import savgoldiff, polydiff, spectraldiff
 
-__all__ = ['lineardiff', 'savgoldiff', 'polydiff', 'spectraldiff'] # So these get treated as direct members of the module by sphinx
+__all__ = ['lineardiff', 'spectraldiff'] # So these get treated as direct members of the module by sphinx
+# polydiff and savgoldiff are still imported for now for backwards compatibility but are not
+# documented as part of this module, since they've moved
