+++ pynumdiff/smooth_finite_difference/__init__.py
@@ -2,4 +2,6 @@
 """
 from ._smooth_finite_difference import mediandiff, meandiff, gaussiandiff, friedrichsdiff, butterdiff, splinediff
 
-__all__ = ['mediandiff', 'meandiff', 'gaussiandiff', 'friedrichsdiff', 'butterdiff', 'splinediff'] # So automodule from the .rst finds them
+__all__ = ['mediandiff', 'meandiff', 'gaussiandiff', 'friedrichsdiff', 'butterdiff'] # So automodule from the .rst finds them
+ # splinediff is still included in the imports list so backwards compatibility isn't broken, but excluded
+ # from the __all__ list so sphinx doesn't try to document it from this module.
